requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
streamlit-option-menu>=0.3.6
Pillow>=10.0.0
twilio>=8.0.0
bcrypt>=4.0.0
python-dotenv>=1.0.0

# Bibliotecas para notificações (opcional - descomentar se for usar)
# twilio>=8.0.0  # Para envio de SMS
# Já incluído no Python: smtplib, email (para envio de emails)

